requirements.txt
groq
langchain
langchain-community
langchain-groq
faiss-cpu
pdfplumber
sentence-transformers
datasets
numpy
langgraph
ragas
streamlit